Update Code and Logic | The bug of backup and restore has not been fixed

+++ backend/table_services.py
@@ -1,7 +1,7 @@
-from flask import render_template, request, redirect, session, url_for
+from flask import jsonify, render_template, request, redirect, session, url_for
 from db_admin import MySQLAdmin
 import re
-
+from key_services import KeyServices
 mysql_admin = MySQLAdmin()
 
 class TableServices:
@@ -32,9 +32,34 @@ class TableServices:
         cursor.execute(f'SELECT * FROM `{dbname}`.`{table}`;')
         rows = cursor.fetchall()
 
+        # FK
+        cursor.execute("""
+                SELECT CONSTRAINT_NAME, COLUMN_NAME,
+                    REFERENCED_TABLE_NAME,
+                    REFERENCED_COLUMN_NAME
+                FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE
+                WHERE TABLE_SCHEMA = %s
+                    AND TABLE_NAME = %s
+                    AND REFERENCED_TABLE_NAME IS NOT NULL
+                """, (dbname, table))
+        foreign_keys = cursor.fetchall()
+
+        # REFERENCE TABLE
+        cursor.execute("""
+                SELECT
+                    TABLE_NAME,
+                    COLUMN_NAME,
+                    CONSTRAINT_NAME
+                FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE
+                WHERE TABLE_SCHEMA = %s
+                    AND REFERENCED_TABLE_NAME = %s
+                """, (dbname, table))
+        referenced_by = cursor.fetchall()
+        cursor.execute("SHOW TABLES FROM `{}`;".format(dbname))
+        tables = [t[0] for t in cursor.fetchall()]
         conn.close()
 
-        return render_template('table_view.html', dbname=dbname, table=table, columns=columns, rows=rows)
+        return render_template('table_view.html', dbname=dbname, table=table, columns=columns, rows=rows, tables=tables, foreign_keys=foreign_keys, referenced_by=referenced_by)
 
 
     # Create Table
@@ -52,7 +77,7 @@ class TableServices:
 
         sql = f"""
         CREATE TABLE `{dbname}`.`{table}` (
-            id NVARCHAR(255) PRIMARY KEY
+            id VARCHAR(255) PRIMARY KEY NOT NULL
         );
         """
 
@@ -76,8 +101,20 @@ class TableServices:
         cursor = conn.cursor()
 
         try:
-            cursor.execute(f"DROP TABLE `{dbname}`.`{table}`;")
-            conn.commit()
+            cursor.execute("""
+                SELECT TABLE_NAME
+                FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE
+                WHERE TABLE_SCHEMA = %s
+                    AND REFERENCED_TABLE_NAME = %s
+            """, (dbname, table))
+            refs = cursor.fetchall()
+            if refs:
+                conn.close()
+                return redirect(url_for("list_tables", dbname=dbname, msg=f"Không thể xóa. Đang được tham chiếu bởi: {[r[0] for r in refs]}"))
+            else:
+                cursor.execute(f"DROP TABLE `{dbname}`.`{table}`;")
+                conn.commit()
+
         except Exception as e:
             return str(e)
 
@@ -85,6 +122,25 @@ class TableServices:
         return redirect(url_for('list_tables', dbname=dbname, msg="Xóa bảng thành công!!!"))
 
     # -------------- COLUMN --------------------
+
+    # GET COLUMNS
+    @staticmethod
+    def get_table_columns(dbname, table):
+        if "username" not in session:
+            return jsonify(columns=[]), 401
+
+        try:
+            conn = mysql_admin.user_conn(session['username'], session['user_pass'])
+            cursor = conn.cursor()
+
+            # Lấy danh sách các cột
+            cursor.execute(f"DESCRIBE `{dbname}`.`{table}`;")
+            cols = [c[0] for c in cursor.fetchall()]
+
+            conn.close()
+            return jsonify(columns=cols)
+        except Exception as e:
+            return jsonify(columns=[], error=str(e)), 500
 
     # Add Column
     @staticmethod
@@ -114,23 +170,47 @@ class TableServices:
 
         try:
             cursor.execute("""
-                        SELECT 1
+                        SELECT CONSTRAINT_NAME
                         FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE
                         WHERE TABLE_SCHEMA = %s
                           AND TABLE_NAME = %s
                           AND COLUMN_NAME = %s
-                          AND CONSTRAINT_NAME = 'PRIMARY'
-                    """, (dbname, table, col))
+                          AND REFERENCED_TABLE_NAME IS NOT NULL
+             """, (dbname, table, col))
+            fk = cursor.fetchone()
+            if fk:
+                cursor.execute(
+                    f"ALTER TABLE `{dbname}`.`{table}` DROP FOREIGN KEY `{fk[0]}`"
+                )
 
-            is_pk = cursor.fetchone()
-            if is_pk:
-                cursor.execute(f"ALTER TABLE `{dbname}`.`{table}` DROP PRIMARY KEY")
+            cursor.execute("""
+                SELECT 1
+                FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE
+                WHERE TABLE_SCHEMA = %s
+                    AND REFERENCED_TABLE_NAME = %s
+                    AND REFERENCED_COLUMN_NAME = %s
+            """, (dbname, table, col))
+            if cursor.fetchone():
+                conn.close()
+                return redirect(f"/database/{dbname}/{table}?err_msg=Cột đang được bảng khác tham chiếu")
+
+            cursor.execute('''
+                SELECT COUNT(*)
+                FROM INFORMATION_SCHEMA.COLUMNS
+                WHERE TABLE_SCHEMA = %s
+                    AND TABLE_NAME = %s
+            ''', (dbname, table))
+            col_count = cursor.fetchone()[0]
+            if col_count <= 1:
+                conn.close()
+                return redirect(f"/database/{dbname}/{table}?err_msg=Không thể xóa cột cuối cùng!!")
 
             cursor.execute(f'ALTER TABLE `{dbname}`.`{table}` DROP COLUMN `{col}`;')
             conn.commit()
 
         except Exception as e:
             conn.rollback()
+            conn.close()
             return redirect(
                 f"/database/{dbname}/{table}?err_msg={str(e)}"
             )
@@ -174,8 +254,16 @@ class TableServices:
         conn = mysql_admin.user_conn(session['username'], session['user_pass'])
         cursor = conn.cursor()
 
+        pk = KeyServices.get_primary_key(cursor, dbname, table)
+        pk_value = request.form.get(pk)
+
         cursor.execute(f'DESCRIBE `{dbname}`.`{table}`;')
         columns = [col[0] for col in cursor.fetchall()]
+
+        if pk_value is None or pk_value.strip() == "":
+            return redirect(
+                f"/database/{dbname}/{table}?err_msg=PRIMARY KEY không được rỗng"
+            )
 
         values = []
         for col in columns:
@@ -199,6 +287,10 @@ class TableServices:
         conn = mysql_admin.user_conn(session['username'], session['user_pass'])
         cursor = conn.cursor()
 
+        pk = KeyServices.get_primary_key(cursor, dbname, table)
+        if not pk:
+            conn.close()
+            return redirect(f"/database/{dbname}/{table}?err_msg=Bảng chưa có PRIMARY KEY nên không thể sửa dòng")
         cursor.execute(f"DESCRIBE `{dbname}`.`{table}`;")
         columns = [col[0] for col in cursor.fetchall()]
 
@@ -211,7 +303,7 @@ class TableServices:
 
         values.append(rowid)
 
-        sql = f"UPDATE `{dbname}`.`{table}` SET {','.join(assignments)} WHERE id=%s"
+        sql = f"UPDATE `{dbname}`.`{table}` SET {','.join(assignments)} WHERE `{pk}`=%s"
         cursor.execute(sql, values)
         conn.commit()
         conn.close()
@@ -224,7 +316,9 @@ class TableServices:
         conn = mysql_admin.user_conn(session['username'], session['user_pass'])
         cursor = conn.cursor()
 
-        cursor.execute(f'DELETE FROM `{dbname}`.`{table}` WHERE id=%s', (rowid,))
+        pk = KeyServices.get_primary_key(cursor, dbname, table)
+
+        cursor.execute(f'DELETE FROM `{dbname}`.`{table}` WHERE `{pk}`=%s', (rowid,))
         conn.commit()
         conn.close()
